gladiator classes in progress
got gladiator entities to display successfully when called. We need to build out their properties and then create game logic tied to buttons in a combat screen. We also need to build out the weapon entities, or hard code weapons as a part of the galdiator for time sake and just figure out an upgrade system.